+++ Hog-Core/src/Hog/Renderer/Buffer.h
@@ -13,15 +13,15 @@ namespace Hog {
 	class Buffer
 	{
 	public:
-		static Ref<Buffer> Create(BufferDescription description, uint32_t size);
+		static Ref<Buffer> Create(BufferDescription description, size_t size);
 	public:
-		Buffer(BufferDescription description, uint32_t size);
+		Buffer(BufferDescription description, size_t size);
 		~Buffer();
 
-		void WriteData(void* data, uint32_t size, uint64_t bufferOffset = 0, uint64_t dataOffset = 0);
-		void ReadData(void* data, uint32_t size, uint64_t bufferOffset = 0, uint64_t dataOffset = 0);
+		void WriteData(void* data, size_t size, size_t bufferOffset = 0, size_t dataOffset = 0);
+		void ReadData(void* data, size_t size, size_t bufferOffset = 0, size_t dataOffset = 0);
 		const VkBuffer& GetHandle() const { return m_Handle; }
-		uint32_t GetSize() const { return m_Size; }
+		size_t GetSize() const { return m_Size; }
 		BufferDescription GetBufferDescription() const { return m_Description; }
 
 		VkDeviceAddress GetBufferDeviceAddress();
@@ -33,23 +33,23 @@ namespace Hog {
 		VmaAllocationInfo m_AllocationInfo;
 
 		BufferDescription m_Description;
-		uint32_t m_Size;
+		size_t m_Size;
 	};
 
 	class BufferRegion
 	{
 	public:
-		static Ref<BufferRegion> Create(Ref<Buffer> buffer, uint64_t offset, uint32_t size);
+		static Ref<BufferRegion> Create(Ref<Buffer> buffer, size_t offset, size_t size);
 	public:
-		BufferRegion(Ref<Buffer> buffer, uint64_t offset, uint32_t size);
+		BufferRegion(Ref<Buffer> buffer, size_t offset, size_t size);
 
-		void WriteData(void* data, uint32_t size, uint64_t bufferOffset = 0, uint64_t dataOffset = 0);
-		void ReadData(void* data, uint32_t size, uint64_t bufferOffset = 0, uint64_t dataOffset = 0);
-		uint32_t GetSize() const { return m_Size; }
-		uint64_t GetOffset() const { return m_Offset; }
+		void WriteData(void* data, size_t size, size_t bufferOffset = 0, size_t dataOffset = 0);
+		void ReadData(void* data, size_t size, size_t bufferOffset = 0, size_t dataOffset = 0);
+		size_t GetSize() const { return m_Size; }
+		size_t GetOffset() const { return m_Offset; }
 	private:
 		Ref<Buffer> m_Buffer;
-		uint64_t m_Offset;
-		uint32_t m_Size;
+		size_t m_Offset;
+		size_t m_Size;
 	};
 }
